requirements.txt
pdfkit==0.6.1
django-bulma

